chore: 캘린더, 검색 페이지 폴더 수정 및 정리
