+++ src/app/search/final/page.tsx
@@ -6,7 +6,7 @@ import BlogSection from '@/components/search/BlogSection'
 import ResumeSection from '@/components/search/ResumeSection'
 import SessionSection from '@/components/search/SessionSection'
 import { useEffect, useState } from 'react'
-import { getSearchList } from '../api/getSearchList'
+import { getSearchList } from '../../../api/search/getSearchList'
 import { useSearchParams } from 'next/navigation'
 import SkeletonCardItem from '@/components/search/SkeletonCard'
 import EmptyLottie from '@/components/common/EmptyLottie'
